#118 [MOD] 모델 홈 상단 연결 커스텀 훅 분리
