+++ src/utils/index.ts
@@ -5,3 +5,4 @@ export * from './debug';
 export * from './chat';
 export * from './myRecruitment';
 export * from './mypage';
+export * from './modelHome';
